mkfs.ubifs: add "-F" option for "free-space fixup"
This adds a superblock flag indicating that "free-space fixup" is needed, an
allows it to be set by the user via the "-F" command-line option.  The first
time the filesystem is mounted, this flag will trigger a one-time re-mapping
all LEBs containing free space.  This fixes problems seen on some NAND flash
when a non-UBIFS-aware flash programmer is used.

Artem: add a bit more help text, tweaked the patch a bit.

Thanks to Ben Gardiner <bengardiner@nanometrics.ca> for help.

Signed-off-by: Matthew L. Creech <mlcreech@gmail.com>
Signed-off-by: Artem Bityutskiy <Artem.Bityutskiy@nokia.com>